refactor(spring-boot): extract FormFields into its own class
Move FormFields and the session-name computation out of JkleeFilesEndpoint
into a dedicated FormFields class with a FormFields.from(prefix, results)
factory. Rename the test to FormFieldsTest.
